Initial setup on Ubuntu 18.10.

+++ build_bpf_modules_opensource.sh
@@ -15,20 +15,34 @@
  # with this program; if not, write to the Free Software Foundation, Inc.,
  # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-set -xeo pipefail
+set -eo pipefail
 # this script must be run inside katran's project root
 # if you are adding new bpf prog:
 # 1) put it into bpf/ dir
 # 2) edit Makefile (add new prog into always += section)
 
-CLANG_PATH="$(pwd)/deps/clang/clang+llvm-7.0.1-x86_64-linux-gnu-ubuntu-18.04"
+#CLANG_PATH="$(pwd)/deps/clang/clang+llvm-7.0.1-x86_64-linux-gnu-ubuntu-18.04"
+CLANG=`which clang`
+if [ "$CLANG" == "" ]; then
+    echo "clang executable not found."
+    exit 1
+fi
+
+LLC=`which llc`
+if [ "LLC" == "" ]; then
+    echo "llc executable not found."
+    exit 1
+fi
+
 rm -rf ./deps/bpfprog
 mkdir -p ./deps/bpfprog/include
+
 cp ./katran/lib/Makefile-bpf ./deps/bpfprog/Makefile
 cp -r ./katran/lib/bpf ./deps/bpfprog/
 cp -r ./katran/decap/bpf ./deps/bpfprog/
 cp ./katran/lib/linux_includes/* ./deps/bpfprog/include/
-cd ./deps/bpfprog && LD_LIBRARY_PATH="${CLANG_PATH}/lib" make \
-  EXTRA_CFLAGS="$*" \
-  LLC="${CLANG_PATH}/bin/llc" CLANG="${CLANG_PATH}/bin/clang"
+
+cd ./deps/bpfprog && make EXTRA_CFLAGS="$*" LLC=$LLC CLANG=$CLANG
+
 echo "BPF BUILD COMPLITED"
+exit 0
